chore: move Auth0 class into its own file
